requirements.txt
openai==0.27.9
discord.py==2.1.1
